Beta

+++ gulpfile.js
@@ -200,8 +200,13 @@ async function unpackPacks() {
 /**
  * Cook db source json files into .db files with nedb
  */
+var cookErrorCount = 0;
+var packErrors = {};
 async function cookPacks() {
     console.log(`Cooking db files`);
+    
+    let compendiumMap = {};
+    let allItems = [];
     
     let sourceDir = "./src/items";
     let directories = await fs.readdirSync(sourceDir);
@@ -211,8 +216,12 @@ async function cookPacks() {
         
         console.log(`Processing ${directory}`);
         
-        console.log(`> Removing ${outputFile}`);
-        await fs.unlinkSync(outputFile);
+        if (fs.existsSync(outputFile)) {
+            console.log(`> Removing ${outputFile}`);
+            await fs.unlinkSync(outputFile);
+        }
+        
+        compendiumMap[directory] = {};
         
         let db = new AsyncNedb({ filename: outputFile, autoload: true });
         
@@ -223,19 +232,124 @@ async function cookPacks() {
             let jsonInput = await fs.readFileSync(filePath);
             jsonInput = JSON.parse(jsonInput);
             
+            compendiumMap[directory][jsonInput._id] = jsonInput;
+            allItems.push({pack: directory, data: jsonInput, file: file});
+            
             await db.asyncInsert(jsonInput);
         }
     }
     
-    console.log(`\nCook finished.\n`);
+    console.log(`\nStarting consistency check.`);
+    
+    cookErrorCount = 0;
+    packErrors = {};
+    
+    for (let item of allItems) {
+        let data = item.data;
+        if (!data || !data.data || !data.data.description) continue;
+        
+        let desc = data.data.description.value;
+        if (!desc) continue;
+        
+        let pack = item.pack;
+        
+        let errors = [];
+        let itemMatch = [...desc.matchAll(/@Item\[([^\]]*)\]{([^}]*)}/gm)];
+        if (itemMatch && itemMatch.length > 0) {
+            for (let localItem of itemMatch) {
+                let localItemId = localItem[1];
+                let localItemName = localItem[2];
+
+                // The current pack must be a valid pack from the compendium map.
+                if (!(pack in compendiumMap)) {
+                    if (!(pack in packErrors)) {
+                        packErrors[pack] = [];
+                    }
+                    packErrors[pack].push(`${item.file}: '${localItemName}' (with id: ${localItemId}) cannot find its own pack '${pack}'.`);
+                    cookErrorCount++;
+                    continue;
+                }
+                
+                // @Item links must link to a valid item ID.
+                if (!(localItemId in compendiumMap[pack])) {
+                    if (!(pack in packErrors)) {
+                        packErrors[pack] = [];
+                    }
+                    packErrors[pack].push(`${item.file}: '${localItemName}' (with id: ${localItemId}) not found in '${pack}'.`);
+                    cookErrorCount++;
+                }
+            }
+        }
+        
+        let compendiumMatch = [...desc.matchAll(/@Compendium\[([^\.]*)\.([^\.]*)\.([^\]]*)\]{([^}]*)}/gm)];
+        if (compendiumMatch && compendiumMatch.length > 0) {
+            for (let otherItem of compendiumMatch) {
+                let system = otherItem[1];
+                let otherPack = otherItem[2];
+                let otherItemId = otherItem[3];
+                let otherItemName = otherItem[4];
+
+                // @Compendium links must link to sfrpg compendiums.
+                if (system !== "sfrpg") {
+                    if (!(pack in packErrors)) {
+                        packErrors[pack] = [];
+                    }
+                    packErrors[pack].push(`${item.file}: Compendium link to '${otherItemName}' (with id: ${otherItemId}) is not referencing the sfrpg system, but instead using '${system}'.`);
+                    cookErrorCount++;
+                }
+                
+                // @Compendium links to the same compendium could be @Item links instead.
+                /*if (otherPack === pack) {
+                    if (!(pack in packErrors)) {
+                        packErrors[pack] = [];
+                    }
+                    packErrors[pack].push(`${item.file}: Compendium link to '${otherItemName}' (with id: ${otherItemId}) is referencing the same compendium, consider using @Item[${otherItemId}] instead.`);
+                    cookErrorCount++;
+                }*/
+                
+                // @Compendium links must link to a valid compendium.
+                if (!(otherPack in compendiumMap)) {
+                    if (!(pack in packErrors)) {
+                        packErrors[pack] = [];
+                    }
+                    packErrors[pack].push(`${item.file}: '${otherItemName}' (with id: ${otherItemId}) cannot find '${pack}', is there an error in the compendium name?`);
+                    cookErrorCount++;
+                    continue;
+                }
+                
+                // @Compendium links must link to a valid item ID.
+                if (!(otherItemId in compendiumMap[otherPack])) {
+                    if (!(pack in packErrors)) {
+                        packErrors[pack] = [];
+                    }
+                    packErrors[pack].push(`${item.file}: '${otherItemName}' (with id: ${otherItemId}) not found in '${otherPack}'.`);
+                    cookErrorCount++;
+                }
+            }
+        }
+    }
+
+    console.log(`\nUpdating items with updated IDs.\n`);
     
     await unpackPacks();
     
+    console.log(`\nCook finished with ${cookErrorCount} errors.\n`);
+
     return 0;
 }
 
 async function postCook() {
-    console.log(`\nCompendiums cooked!\nDon't forget to restart Foundry to refresh compendium data!\n`);
+    
+    if (Object.keys(packErrors).length > 0) {
+        for (let pack of Object.keys(packErrors)) {
+            console.log(`\n${packErrors[pack].length} Errors cooking ${pack}.db:`);
+            for (let error of packErrors[pack]) {
+                console.log(`> ${error}`);
+            }
+        }
+    }
+    
+    console.log(`\nCompendiums cooked with ${cookErrorCount} errors!\nDon't forget to restart Foundry to refresh compendium data!\n`);
     return 0;
 }
